noteezy-backend/app/models.py
from app import db
import datetime
import jwt
import os

class User(db.Model):
    id = db.Column(db.Integer, primary_key=True)
    email = db.Column(db.String(255), unique=True, nullable=False)
    password = db.Column(db.String(255), nullable=False)
    plan = db.Column(db.String(10), default="free")
    monthly_notes = db.Column(db.Integer, default=0)

    def generate_token(self):
        return jwt.encode({"user_id": self.id}, os.getenv("JWT_SECRET"), algorithm="HS256")

    @staticmethod
    def verify_token(token):
        try:
            payload = jwt.decode(token, os.getenv("JWT_SECRET"), algorithms=["HS256"])
            return User.query.get(payload["user_id"])
        except:
            return None

class Note(db.Model):
    id = db.Column(db.Integer, primary_key=True)
    user_id = db.Column(db.Integer, db.ForeignKey("user.id"))
    content = db.Column(db.Text)
    created_at = db.Column(db.DateTime, default=datetime.datetime.utcnow)