noteezy-backend/requirements.txt

flask
flask-cors
flask-sqlalchemy
flask-limiter
transformers
torch
Pillow
fpdf
langdetect
python-dotenv
